Fixed warnings.

+++ phoneProfilesPlus/src/main/java/sk/henrichg/phoneprofilesplus/BrightnessDialogPreference.java
@@ -394,7 +394,7 @@ public class BrightnessDialogPreference extends
         // Round the value to the closest integer value.
         //noinspection ConstantConditions
         if (stepSize >= 1) {
-            value = Math.round(newValue/stepSize)*stepSize;
+            value = Math.round((float)newValue/stepSize)*stepSize;
         }
         else {
             value = newValue;
